rename WorkflowStepButton to EvaluationStepButton
